+++ VRCVideoCacher/Program.cs
@@ -14,7 +14,7 @@ namespace VRCVideoCacher;
 internal sealed class Program
 {
     public static string YtdlpHash = string.Empty;
-    public const string Version = "2026.1.20";
+    public const string Version = "2026.1.21";
     public static readonly ILogger Logger = Log.ForContext("SourceContext", "Core");
     public static readonly string CurrentProcessPath = Path.GetDirectoryName(Environment.ProcessPath) ?? string.Empty;
     public static readonly string DataPath = OperatingSystem.IsWindows()
@@ -44,7 +44,7 @@ internal sealed class Program
         Log.Logger = new LoggerConfiguration()
             .MinimumLevel.Debug()
             .WriteTo.Console(new ExpressionTemplate(
-                "[{@t:HH:mm:ss} {@l:u3} {Coalesce(Substring(SourceContext, LastIndexOf(SourceContext, '.') + 1),'<none>')}] {@m}\n{@x}",
+                "[{@t:HH:mm:ss} {@l:u3} {Coalesce(Substring(SourceContext, LastIndexOf(SourceContext, '.') + 1),'<none>')}] {@m}\n\r{@x}",
                 theme: TemplateTheme.Literate))
             .WriteTo.Sink(new UiLogSink())
             .CreateLogger();